+++ CHANGES.md
@@ -1,5 +1,11 @@
 # **Diagnosticism.Python** Changes
 
+## 0.11.0 - 17th July 2025
+
+* ~ `enable_logging()` now also takes 2-parameters (environment variable name, default value);
+* ~ `enable_tracing()` now also takes 2-parameters (environment variable name, default value);
+
+
 ## 0.10.1 - 17th July 2025
 
 * ~ fixed package error;